Install sshd keys with correct permissions
On a macOS host, a guest's virtiofs mount's files will always report a
UID and GID that matches the EUID and EGID. That is: any user can access
all files in a virtiofs filesystem.  It is impossible to keep their
contents private.

Use a new `sshdkeys.service` to temporarily mount, copy (with correct
owner/group/permissions), and unmount the sensitive key material before
`sshd` is started so users cannot gain access to the material.