Split `hostname` and `user` into specific values
Use as many different names for logically-distinct values as possible.
The intent is to have them be "stringly-typed" values so an accidental
replacement of one with another will raise an error and be noticed.

The only places where this isn't done is when a name escapes out to be
user-visible (eg: part of the user-facing "API").  Leave those to be the
same "rosetta-builder" value to preserve a good user experience.