Recreate the VM if its image changes
Fix changing configuration options that require changes to the VM image
(like changing `onDemand`).  Previously the VM would always keep the image it generated the first time it was run.  If this got out of sync with the current settings then it would often fail to start and builds would fail as they wouldn't be able to reach the VM over SSH.

Since this deletes (and recreates) any existing VM when there's a new
image, the cached contents of the previous VM's Nix store wlil be lost.