Add `allowNonRootSshAccess` module option
Allow regular, non-root users SSH access to the VM via
`ssh rosetta-builder` (without `sudo`).

This is not made the default because restricting access provides a bit
of defense in depth.  Helping to protect against, for example, some
kinds of NixOS vulnerabilities.

It *shouldn't* be possible to use this feature to compromise the
integrity of builds that the host remotes to the VM.

Closes: https://github.com/cpick/nix-rosetta-builder/issues/18